Preferneces API endpoint implemented and all longs changed to ints

+++ backend/src/main/java/com/mealchemy/auth/dto/AuthResponse.java
@@ -4,7 +4,7 @@ package com.mealchemy.auth.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record AuthResponse( //records are immutable and auto generate constructors
-    @JsonProperty("user_id") Long userId,
+    @JsonProperty("user_id") Integer userId,
     @JsonProperty("token") String accessToken,
     @JsonProperty("onboarding_required") boolean onboardingRequired
 ) {}
