+++ backend/src/main/java/com/mealchemy/auth/model/Role.java
@@ -15,7 +15,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private Long roleId;
+    private Integer roleId;
 
     @Column(name = "role_name", nullable = false, unique = true)
     private String roleName;
@@ -25,7 +25,7 @@ public class Role {
     @Column(name = "permissions", columnDefinition = "jsonb")
     private Map<String, Object> permissions;
 
-    public Long getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
