+++ backend/src/main/java/com/mealchemy/auth/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> { //interface never gets implemented - params = this repository manages User objects and their primary key type is a Long
+public interface UserRepository extends JpaRepository<User, Integer> { //interface never gets implemented - params = this repository manages User objects and their primary key type is a Integer
     // custom methods
     Optional<User> findByEmail(String email);
     boolean existsByEmail(String email);
